% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_eval.R
\name{dev_eval}
\alias{dev_eval}
\title{Extract explained deviance from BRT}
\usage{
dev_eval(x)
}
\arguments{
\item{x}{A fitted BRT model object, e.g. as returned from \code{\link{brt_sdm}}}
}
\value{
The deviance
}
\description{
function to extract explained deviance from BRT
}
\examples{
# use defaults
sim1 <- SimulateWorld(start.year=2015, n.year=20)
fit <- brt_sdm(sim, "temp")
dev_eval(fit)

}
