% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_sdm.R
\name{gam_sdm}
\alias{gam_sdm}
\title{SDM fit with gam}
\usage{
gam_sdm(x, covariates = NULL, start.forecast.year = 2021, k = NULL)
}
\arguments{
\item{x}{(required) An operating model as output from one of the operating model functions (such as \code{sim <- \link{SimulateWorld}()} OR list with meta$abund_enviro and grid from the operating model (\code{sim$grid}).}

\item{covariates}{Covariates to use in the SDM. Must be in the operating model output (in the columns of x$grid). If left off, all covariates in x (in x$meta$covariates) are used.}

\item{start.forecast.year}{The years less will be used for fitting and the years greater than are the forecasted years.}

\item{k}{The smoothness of the fit can be restricted by passing in k. If left off, the default will be used.}
}
\value{
A \link[=SDM_class]{SDM} object, which is a list with the presence and abundance fits and the meta data.
}
\description{
Build a gam species distribution model from the output from an operating model. 
Models fit to both the
presence (pres column) and log(abundance) are returned.  k can be passed
in to change the smoothing.
}
\examples{
# use defaults
sim <- SimulateWorld(n.year=10)
fit <- gam_sdm(sim, "temp")
# plot center of gravity
\dontrun{ #plot center of gravity (slow)
plot_cog(sim, fit)
}

# presence fit
summary(fit$presence)
# abundance fit
summary(fit$abundance)

}
