% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_sdm.R
\name{mlp_sdm}
\alias{mlp_sdm}
\title{SDM fit with MLP}
\usage{
mlp_sdm(
  x,
  covariates,
  response = c("pres", "abundance"),
  start.forecast.year = 2020
)
}
\arguments{
\item{x}{(required) An operating model as output from one of the operating model functions (such as \code{sim <- \link{SimulateWorld}()} OR just the grid from the operating model (\code{sim$grid}).}

\item{covariates}{(required) covariates to use in the SDM. Must be in the operating model output (in the columns of x$grid).}

\item{start.forecast.year}{The years equal or less will be used for fitting and the years greater than are the forecasted years.}
}
\description{
Build a MLP species distribution model from the output from a operating model (or grid) using the neuralnet package. The model is fit with 3 neurons in the hidden layer (which is good for this particular simulation). If it gets much more complicated, can try different values. But training time increases exponentially with additional neurons. The variables are normalized to a common scale before training the MLPs.
}
\examples{
sim <- SimulateWorld()
fit <- mlp_sdm(sim, "temp")
dat <- fit$data #data used for fit
dat$mlpAbun <- predict(fit, dat)
plot(dat$temp, dat$mlpAbun)

}
