% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OM.R
\name{plot.OM}
\alias{plot.OM}
\title{Plots OM object}
\usage{
\method{plot}{OM}(x, ..., start.forecast.year = NULL)
}
\arguments{
\item{x}{An \code{\link[=OM_class]{OM}} object output by one of the simulation functions.}

\item{...}{Not used.}

\item{start.forecast.year}{If passed in, the training data will be blue and the test data will be grey. Otherwise the yearly averages will be in black.}
}
\description{
Plots the yearly averages of the values in the OM object.
}
