% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_POM.R
\name{plot.POM}
\alias{plot.POM}
\title{Plots POM object}
\usage{
\method{plot}{POM}(x, ...)
}
\arguments{
\item{x}{An \code{\link[=POM_class]{POM}} object output by \code{\link{predict.OM}}.}

\item{...}{Not used.}
}
\description{
Plots the yearly averages of the values in the POM object.
}
