% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SDM.R
\name{plot.SDM}
\alias{plot.SDM}
\title{Plots OM object}
\usage{
\method{plot}{SDM}(x, type = c("cog", "abund.by.year"), ...)
}
\arguments{
\item{x}{SDM object}

\item{type}{type of plot}
}
\description{
plots a SDM object
}
