% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cog.R
\name{plot_cog}
\alias{plot_cog}
\title{Plots COG from SDM}
\usage{
plot_cog(x, ...)
}
\arguments{
\item{x}{An \code{\link[=OMclass]{OM}} object output by one of the simulation functions.}

\item{...}{SDMs to compare. Must be an \link[=SDM_class]{SDM} object output from one of the fitting functions (\code{\link{gam_sdm}},
\code{\link{brt_sdm}}, or \code{\link{mlp_sdm}}) or \link[=POM_class]{POM} object output by \code{\link{predict.OM}}.}
}
\description{
plots the center of gravity versus time
}
\details{
The SDMs which will be compared are \link[=SDM_class]{SDM} objects and
 have the `start.forecast.year` as part of their meta data (in $meta).
 The COG will be plotted for the training data (to which the SDM was fit) 
 and the test data (after `start.forecast.year`). If `start.forecast.year`
 is greater than the last year of data, then the forecast years are not
 plotted (since there is not data for those years).
}
\examples{
sim <- SimulateWorld(start.year=2015, n.year=20)
mlp.fit <- mlp_sdm(sim, "temp")
plot_cog(sim, mlp.fit)

}
