% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POM-class.R
\docType{class}
\name{POM_class}
\alias{POM_class}
\title{POM class}
\description{
This is the class for output from the `predict.OM` function (\code{\link{predict.OM}}). The POM object is a data from with the predicted presence (pred.a), abundance (pred.a, logged if model is log-normal), and expected abundance (pred = pred.p x exp(pred.a), if log-normal).
This allows easy plotting from predictions.
}
