% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_SDM.R
\name{predict.SDM}
\alias{predict.SDM}
\title{Predict for SDM object}
\usage{
\method{predict}{SDM}(
  object,
  newdata = NULL,
  start.forecast.year = 2021,
  alpha = 0.6826895,
  silent = FALSE,
  ...
)
}
\arguments{
\item{object}{SDM object}

\item{newdata}{(optional) Optional new data frame over which to do the prediction.}

\item{start.forecast.year}{The years equal or less will be used for fitting and the years greater than are the forecasted years.}

\item{alpha}{The CIs alpha. Default is 1 sigma (alpha = 68\%)}

\item{silent}{No output printed.}

\item{...}{Not used.}
}
\value{
Invisibly returns a data frames with the prediction. If newdata is not passed in, then the hindcast and forecast prediction is returned. The "type" column has "hindcast" and "forecast" to indicate which is which. This data frame is the same shape as object$grid except the extra column "type". The first 6 rows of the prediction are printed unless silent=TRUE. If "newdata" is passed in then the prediction for the new data only is returned. "newdata" must be a data frame with the required covariates.
}
\description{
Does a hindcast and forecast with SDM fit (sdm returned from one of the sdm
fitting functions). If abund_enviro is log-normal, then 
 a presence model and an abundance model are fit. If 
 abund_enviro is poisson, then only the abundance model is fit. Both a hindcast
 and a forecast will be returned. If model is "gam", then 95% CIs will be returned also.
}
