% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_OM.R
\name{print.OM}
\alias{print.OM}
\title{Prints OM object}
\usage{
\method{print}{OM}(x, ...)
}
\arguments{
\item{x}{OM object}

\item{...}{Not used}
}
\description{
prints a \link[=OMclass]{OM} object returned from one of the simulation functions.
}
