% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_SDM.R
\name{print.SDM}
\alias{print.SDM}
\title{Prints SDM object}
\usage{
\method{print}{SDM}(x, ...)
}
\arguments{
\item{x}{SDM object}

\item{...}{Not used}
}
\description{
prints a \link[=SDM_class]{SDM} object returned from one of the simulation functions.
}
