% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM-class.R
\docType{class}
\name{SDM_class}
\alias{SDM_class}
\title{SDM class}
\description{
This is the class for output from the fitting functions. The output is the list 
normally returned by the fitting function with meta from the OM objected added.
This allows easy plotting from fitted SDMs.  See \code{\link[=OMclass]{OM}} for
 a description of the meta element.
}
