% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{WRAP-package}
\alias{WRAP}
\alias{WRAP-package}
\title{WRAP Package}
\description{
This package creates simulated presence/absence and abundance fields and
fits SDMs.
}
\author{
\strong{Maintainer}: Stephanie Brodie \email{stephanie.brodie@noaa.gov}

Authors:
\itemize{
  \item James Smith \email{james.smith@noaa.gov}
  \item Eli Holmes \email{eli.holmes@noaa.gov}
}

}
\keyword{internal}
