% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brt_sdm.R
\name{brt_sdm}
\alias{brt_sdm}
\title{SDM fit with BRT}
\usage{
brt_sdm(
  x,
  covariates = NULL,
  response = c("pres", "abundance"),
  start.forecast.year = 2021
)
}
\arguments{
\item{x}{(required) An operating model as output from one of the operating model functions (such as \code{sim <- \link{SimulateWorld}()} OR list with meta$abund_enviro and grid from the operating model (\code{sim$grid}).}

\item{covariates}{Covariates to use in the SDM. Must be in the operating model output (in the columns of x$grid). If left off, all covariates in x (in x$meta$covariates) are used.}

\item{start.forecast.year}{The years less will be used for fitting and the years greater than are the forecasted years.}
}
\description{
Build a boosted regression tree species distribution model from the output from a operating model using the dismo package.
}
\examples{
sim <- SimulateWorld()
fit <- brt_sdm(sim, "temp")
dev_eval(fit)
plot(fit)

# you can also pass in the grid from sim
# this allows you to modify the grid (add 0s or error)
bad.grid <- sim$grid
bad.grid$abundance <- bad.grid$abundance + rnorm(nrow(bad.grid),0,0.05)
fit <- brt_sdm(bad.grid, "temp", response="abundance")

}
