% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brt_sdm.R
\name{brt_sdm}
\alias{brt_sdm}
\title{SDM fit with BRT}
\usage{
brt_sdm(
  x,
  covariates = NULL,
  start.forecast.year = 2021,
  control = list(tree.complexity = 3, learning.rate = 0.01, bag.fraction = 0.6)
)
}
\arguments{
\item{x}{(required) An operating model as output from one of the operating model functions (such as \code{sim <- \link{SimulateWorld}()} OR list with meta$abund_enviro and grid from the operating model (\code{sim$grid}).}

\item{covariates}{Covariates to use in the SDM. Must be in the operating model output (in the columns of x$grid). If left off, all covariates in x (in x$meta$covariates) are used.}

\item{start.forecast.year}{The years less will be used for fitting and the years greater than are the forecasted years.}

\item{control}{Parameters for the `dismo::gbm.step` call.}
}
\value{
A list with the presence and abundance fits and the meta data.
}
\description{
Build a boosted regression tree species distribution model from the output from a operating model using the dismo package. 
Models fit to both the
presence (pres column) and log(abundance) are returned.
}
\examples{
sim <- SimulateWorld()
# abundance fit
fit <- brt_sdm(sim, "temp")$abundance
dev_eval(fit)
plot(fit)

# you can also pass in the grid from sim
# this allows you to modify the grid (add 0s or error)
bad.grid <- sim$grid
bad.grid$abundance <- bad.grid$abundance + rnorm(nrow(bad.grid),0,0.05)
fit <- brt_sdm(bad.grid, "temp")$abundance

}
