% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_sdm.R
\name{gam_sdm}
\alias{gam_sdm}
\title{SDM fit with gam}
\usage{
gam_sdm(
  x,
  covariates,
  response = c("pres", "abundance"),
  start.forecast.year = 2020,
  k = NULL
)
}
\arguments{
\item{x}{(required) An operating model as output from one of the operating model functions (such as \code{sim <- \link{SimulateWorld}()} OR just the grid from the operating model (\code{sim$grid}).}

\item{covariates}{(required) covariates to use in the SDM. Must be in the operating model output (in the columns of x$grid).}

\item{response}{"pres" or "abundance"}

\item{start.forecast.year}{The years equal or less will be used for fitting and the years greater than are the forecasted years.}

\item{k}{The smoothness of the fit can be restricted by passing in k. If left off, the default will be used.}
}
\description{
Build a gam species distribution model from the output from a operating model.
}
\examples{
# use defaults
sim <- SimulateWorld()
fit <- gam_sdm(sim, "temp")
summary(fit)
plot(fit)

sim <- SimulateWorld_ROMS()
fit <- gam_sdm(sim, "sst", response="abundance", k=4)
summary(fit)
plot(fit)

}
