% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abund.R
\name{plot_abund}
\alias{plot_abund}
\title{Plot the total yearly abundance}
\usage{
plot_abund(x, ...)
}
\arguments{
\item{x}{Object to plot. Can be a \link[=OM_class]{OM} or \link[=POM_class]{POM} object.}

\item{...}{more objects to compare. Can be a \link[=SDM_class]{SDM}, \link[=POM_class]{POM} or \link[=OM_class]{OM}.}
}
\description{
Plots the total abundance versus year for objects of class OM (returned by simulations functions), SDM (returned by fitting functions) and/or POM (returned by predict function).
}
\details{
If any of the objects to be compared are \link[=SDM_class]{SDM} objects, then x must be a
OM object and predicts will be made for the grid points in the OM object.
If the objects to be compared are \link[=SDM_class]{SDM} objects and
 have the `start.forecast.year` as part of their meta data (in $meta), a reference 
 line will be plotted showing the training versus test years.
 If `start.forecast.year`
 is greater than the last year of data, then no reference line will be plotted.
}
\examples{
sim1 <- SimulateWorld(start.year=2015, n.year=20)
sim2 <- SimulateWorld(start.year=2040, n.year=20)
mlp.fit <- mlp_sdm(sim1, "temp")
plot_abund(sim1, mlp.fit, sim2)

}
