% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abund.R
\name{plot_abund}
\alias{plot_abund}
\title{Plots abundance from SDM}
\usage{
plot_abund(x, ...)
}
\arguments{
\item{x}{An \code{\link[=OMclass]{OM}} object output by one of the simulation functions.}

\item{...}{SDMs to compare. Must be an \link[=SDM_class]{SDM} object output from one of the fitting functions:
\code{\link{gam_sdm}}, \code{\link{brt_sdm}}, or \code{\link{mlp_sdm}}.}
}
\description{
plots the total abundance versus time
}
\examples{
\dontrun{
sim <- SimulateWorld()
mlp.fit <- mlp_sdm(sim, "temp")
plot_abund(sim, mlp.fit)
}

}
