% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cog.R
\name{plot_cog}
\alias{plot_cog}
\title{Plots COG from SDM}
\usage{
plot_cog(x, ...)
}
\arguments{
\item{x}{SDM object}

\item{...}{Additional SDM objects to compare}
}
\description{
plots the center of gravity versus time
}
