% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM-class.R
\docType{class}
\name{SDM_class}
\alias{SDM_class}
\title{SDM class}
\description{
This is the class for output from the fitting functions (\code{\link{gam_sdm}}, \code{\link{mlp_sdm}}, and \code{\link{brt_sdm}}). The SDM object is a
 list with the abundance and presence fits plus meta added from the OM object.
This allows easy plotting from fitted SDMs.  See \code{\link[=OM_class]{OM}} for
 a description of the meta element.
}
