% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatedWorld_ROMS_Function.R
\name{SimulateWorld_ROMS}
\alias{SimulateWorld_ROMS}
\title{Simulated World ROMS temperature}
\usage{
SimulateWorld_ROMS(
  PA_shape = c("logistic", "logistic_prev", "linear"),
  abund_enviro = c("lnorm_low", "lnorm_high", "poisson"),
  dir = file.path(here::here(), "Rasters_2d_monthly")
)
}
\arguments{
\item{PA_shape}{specifies how enviro suitability determines species presence-absence. takes values of "logistic" (SB original), "logistic_prev" (JS, reduces knife-edge), "linear" (JS, reduces knife edge, encourages more absences)}

\item{abund_enviro}{specifies abundance if present, can be "lnorm_low" (SB original), "lnorm_high" (EW), or "poisson" (JS, increases abundance range)}

\item{dir}{}
}
\description{
Function uses temperature data from ROMS to generate species distribution and abundance.
}
\details{
The ROMS data are assumed to be in the working directory in a folder called 'Rasters_2d_monthly/gfdl/sst_monthly'. If the ROMS data folder is different, you can pass that in via the 'dir' argument. However the ROMS data folder must have subfolders 'gfdl/sst_monthly' so that 'paste0(dir,"/gfdl/sst_monthly")' will find the SST ROMS data.
Caution: this function uses the downscaled GCMs to simulate species distrubtion. It has slight differences to the other SimulatedWorld_function that 'randomly' generates environmental data. Remember: 1980-2010 are not observed data (by design)
}
\examples{
test <- SimulatedWorld_ROMS(PA_shape="logistic", abund_enviro="lnorm_low")
head(test)
tail(test)
plot(aggregate(suitability~year, data=test, FUN=mean),type='b')
plot(aggregate(presabs~year, data=test, FUN=sum),type='b')
plot(aggregate(abundance~year, data=test, FUN=sum),type='b')
plot(aggregate(sst~year, data=test, FUN=mean),type='b')

}
