% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatedWorld_ROMS_TrophicInteraction.R
\name{SimulateWorld_ROMS_TrophicInteraction}
\alias{SimulateWorld_ROMS_TrophicInteraction}
\title{Simulated World ROMS Trophic Interactions}
\usage{
SimulateWorld_ROMS_TrophicInteraction(
  PA_shape = c("logistic", "logistic_prev", "linear"),
  abund_enviro = c("lnorm_low", "lnorm_high", "poisson"),
  dir = file.path(here::here(), "Rasters_2d_monthly")
)
}
\arguments{
\item{PA_shape}{specifies how enviro suitability determines species presence-absence. takes values of "logistic" (SB original), "logistic_prev" (JS, reduces knife-edge), "linear" (JS, reduces knife edge, encourages more absences)}

\item{abund_enviro}{specifies abundance if present, can be "lnorm_low" (SB original), "lnorm_high" (EW), or "poisson" (JS, increases abundance range)}

\item{dir}{(optional) The path to the directory where the folder 'gfdl' is.}
}
\value{
Returns the grid for species B. This is an object of class \code{\link[=OMclass]{OM}}, which is a list with "grid" and "meta". "meta" has all the information about the simulation including all the parameters passed into the function.
}
\description{
This code replicates SimulateWorld_ROMS_function, but adds in a trophic interaction, where:
\itemize{
 \item Species A: distribution and abundance driven by SST and Chl-a
 \item Species B: distribution and abundance drivn by SST and Species A
 \item EM for Species B: only have chl-a and temp as covariates. 
}
Note only using GFDL for now
}
\details{
The ROMS data are assumed to be in the working directory in a folder called
 'Rasters_2d_monthly/gfdl/sst_monthly'. If the ROMS data folder is different,
  you can pass that in via the 'dir' argument. However the ROMS data folder
   must have subfolders 'gfdl/sst_monthly' so that 
   'paste0(dir,"/gfdl/sst_monthly")' will find the SST ROMS data.
}
\examples{
\dontrun{
test <- SimulateWorld_ROMS_TrophicInteraction(PA_shape="logistic", abund_enviro="lnorm_low")
}

}
